+++ crates/rust-client/src/keystore/web_keystore.rs
@@ -1,14 +1,17 @@
-use alloc::string::ToString;
-use alloc::sync::Arc;
-use miden_objects::{
-    account::{AccountDelta, AuthSecretKey},
-    Digest, Felt, Word,
-};
-use miden_tx::{auth::TransactionAuthenticator, AuthenticationError};
+use alloc::{string::ToString, sync::Arc, vec::Vec};
+
+use miden_lib::utils::{Deserializable, Serializable};
 use rand::Rng;
-use miden_tx::utils::sync::RwLock;
-use super::{KeyStore, KeyStoreError};
-use crate::store::web_store::account::utils::{get_account_auth_by_pub_key, insert_account_auth};
+
+use super::KeyStoreError;
+use crate::{
+    AuthenticationError, Felt, Word,
+    account::AccountDelta,
+    auth::{AuthSecretKey, TransactionAuthenticator},
+    crypto::Digest,
+    store::web_store::account::utils::{get_account_auth_by_pub_key, insert_account_auth},
+    utils::RwLock,
+};
 
 /// A web-based keystore that stores keys in [browser's local storage](https://developer.mozilla.org/en-US/docs/Web/API/Web_Storage_API)
 /// and provides transaction authentication functionality.
@@ -21,16 +24,10 @@ pub struct WebKeyStore<R: Rng> {
 impl<R: Rng> WebKeyStore<R> {
     /// Creates a new instance of the web keystore with the provided RNG.
     pub fn new(rng: R) -> Self {
-        WebKeyStore {
-            rng: Arc::new(RwLock::new(rng)),
-        }
+        WebKeyStore { rng: Arc::new(RwLock::new(rng)) }
     }
-}
 
-impl<R: Rng> KeyStore for WebKeyStore<R> {
-    fn add_key(&self, key: &AuthSecretKey) -> Result<(), KeyStoreError> {
-        let window = web_sys::window()
-            .ok_or_else(|| KeyStoreError::StorageError("Window not available".to_string()))?;
+    pub async fn add_key(&self, key: &AuthSecretKey) -> Result<(), KeyStoreError> {
         let pub_key = match &key {
             AuthSecretKey::RpoFalcon512(k) => Digest::from(Word::from(k.public_key())).to_hex(),
         };
@@ -43,7 +40,7 @@ impl<R: Rng> KeyStore for WebKeyStore<R> {
         Ok(())
     }
 
-    fn get_key(&self, pub_key: Word) -> Result<Option<AuthSecretKey>, KeyStoreError> {
+    pub fn get_key(&self, pub_key: Word) -> Result<Option<AuthSecretKey>, KeyStoreError> {
         let pub_key_str = Digest::from(pub_key).to_hex();
         let secret_key_hex = get_account_auth_by_pub_key(pub_key_str).map_err(|_| {
             KeyStoreError::StorageError("Failed to get item from local storage".to_string())
@@ -76,29 +73,9 @@ impl<R: Rng> TransactionAuthenticator for WebKeyStore<R> {
         _account_delta: &AccountDelta,
     ) -> Result<Vec<Felt>, AuthenticationError> {
         let mut rng = self.rng.write();
-        let secret_key = self.get_key(pub_key).map_err(|err| AuthenticationError::other(err.to_string()))?;
-        let AuthSecretKey::RpoFalcon512(k) = secret_key
-            .ok_or(AuthenticationError::UnknownPublicKey(Digest::from(pub_key).into()))?;
-        miden_tx::auth::signatures::get_falcon_signature(&k, message, &mut *rng)
-    }
-}
-
-impl<R: Rng> TransactionAuthenticator for WebKeyStore<R> {
-    /// Gets a signature over a message, given a public key.
-    ///
-    /// The public key should correspond to one of the keys tracked by the keystore.
-    ///
-    /// # Errors
-    /// If the public key isn't found in the store, [`AuthenticationError::UnknownPublicKey`] is
-    /// returned.
-    fn get_signature(
-        &self,
-        pub_key: Word,
-        message: Word,
-        _account_delta: &AccountDelta,
-    ) -> Result<Vec<Felt>, AuthenticationError> {
-        let mut rng = self.rng.write();
-        let secret_key = self.get_key(pub_key).map_err(|err| AuthenticationError::other(err.to_string()))?;
+        let secret_key = self
+            .get_key(pub_key)
+            .map_err(|err| AuthenticationError::other(err.to_string()))?;
         let AuthSecretKey::RpoFalcon512(k) = secret_key
             .ok_or(AuthenticationError::UnknownPublicKey(Digest::from(pub_key).into()))?;
         miden_tx::auth::signatures::get_falcon_signature(&k, message, &mut *rng)
