+++ crates/rust-client/src/tests.rs
@@ -31,6 +31,7 @@ use miden_tx::utils::{Deserializable, Serializable};
 use rand::rngs::StdRng;
 
 use crate::{
+    Client, ClientError,
     keystore::FilesystemKeyStore,
     mock::create_test_client,
     rpc::NodeRpcClient,
@@ -49,7 +50,7 @@ async fn insert_new_wallet<R: FeltRng>(
     let key_pair = SecretKey::with_rng(&mut client.rng);
     let pub_key = key_pair.public_key();
 
-    keystore.add_key(&AuthSecretKey::RpoFalcon512(key_pair)).await.unwrap();
+    keystore.add_key(&AuthSecretKey::RpoFalcon512(key_pair)).unwrap();
 
     let mut init_seed = [0u8; 32];
     client.rng.fill_bytes(&mut init_seed);
@@ -78,7 +79,7 @@ async fn insert_new_fungible_faucet<R: FeltRng>(
     let key_pair = SecretKey::with_rng(&mut client.rng);
     let pub_key = key_pair.public_key();
 
-    keystore.add_key(&AuthSecretKey::RpoFalcon512(key_pair)).await.unwrap();
+    keystore.add_key(&AuthSecretKey::RpoFalcon512(key_pair)).unwrap();
 
     // we need to use an initial seed to create the wallet account
     let mut init_seed = [0u8; 32];
